Refactor user role retrieval and improve navigation
This commit introduces the `getRole` utility function for more efficient user role retrieval and simplifies related logic in `app/page.tsx` and `components/Nav.tsx`. Additionally, an admin dashboard link is now conditionally displayed in the navigation bar for users with 'owner' or 'barista' roles.
