+++ app/page.tsx
@@ -1,29 +1,16 @@
 import Image from "next/image";
 import logo from "@/public/logo.png"
 import AuthButton from "@/components/AuthButton";
-import { createClient } from "@/utils/supabase/server";
-import { redirect } from "next/navigation";
-import { jwtDecode } from 'jwt-decode'
-import { log } from "console";
+import {redirect} from "next/navigation";
+import getRole from "@/utils/getRole";
 
 export default async function Index() {
-	const supabase = createClient();
-	const session = await supabase.auth.getSession()
-	const access_token = session.data.session?.access_token
-	if (access_token) {
-		const jwt: any = jwtDecode(access_token as string)
-		const user_role = jwt.user_role // owner, barista, client or null
-		console.log(jwt.sub);
-
-		console.log(user_role);
-
-		// TODO: simplify logic
-		if (user_role === 'client' || user_role === null) {
-			return redirect('/client')
-		}
-		if (user_role === 'owner' || user_role === 'barista') {
-			return redirect('/admin')
-		}
+	const userRole = await getRole()
+	if (userRole === 'client' || userRole === null) {
+		return redirect('/client')
+	}
+	if (userRole === 'owner' || userRole === 'barista') {
+		return redirect('/admin')
 	}
